PHAS/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
pymysql==1.1.0
email-validator==2.1.0
Flask-Mail==0.9.1
Werkzeug==2.3.7
Jinja2==3.1.3
itsdangerous==2.1.2
python-dotenv==1.0.1
